+++ src/main/java/business_layer/Withholding.java
@@ -9,42 +9,41 @@ package business_layer;
  * @author drewm
  */
 public class Withholding {
-    public double id;
+//    Attributes for witholding class
+    public int id;
     public String description;
     public double rate;
-
-
-
-    public Withholding(double id, String description, double rate) {
+//    Constructor for withholding class
+    public Withholding(int id, String description, double rate) {
         this.id = id;
         this.description = description;
         this.rate = rate;
     }
-
+// Getter for id
     public double getId() {
         return id;
     }
-
-    public void setId(double id) {
+// Setter for id
+    public void setId(int id) {
         this.id = id;
     }
-
+// Getter for description
     public String getDescription() {
         return description;
     }
-
+// Setter for description
     public void setDescription(String description) {
         this.description = description;
     }
-
+// Getter for rate
     public double getRate() {
         return rate;
     }
-
+// Setter for rate
     public void setRate(int rate) {
         this.rate = rate;
     }
-
+// To string for Withholding class
     @Override
     public String toString() {
         return "Withholding{" + "id=" + id + ", description=" + description + ", rate=" + rate + '}';
